# openapi.yaml
openapi: 3.1.1
info:
  title: API de Registro de Libros
  version: 1.1.0
  description: >
    API FIRST para registrar, consultar y actualizar libros usando validación automática.

servers:
  - url: http://localhost:3000

paths:
  /libros:
    post:
      summary: Registrar un nuevo libro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibroInput"
      responses:
        "201":
          description: Libro registrado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Libro"
        "400":
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /libros/{id}:
    get:
      summary: Obtener libro por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Libro encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Libro"
        "400":
          description: ID inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Libro no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      summary: Actualizar libro por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibroInput"
      responses:
        "200":
          description: Libro actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Libro"
        "400":
          description: ID inválido o datos incorrectos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Libro no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    LibroInput:
      type: object
      required:
        - titulo
        - clave
        - autor
      properties:
        titulo:
          type: string
          minLength: 5
        clave:
          type: string
          minLength: 8
          maxLength: 8
        autor:
          type: string
          minLength: 3

    Libro:
      type: object
      properties:
        id:
          type: integer
        titulo:
          type: string
        clave:
          type: string
        autor:
          type: string

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object